Basic lookahead bugging.
When follow a wall, the robot will try to round the corner, and otherwise it uses the lookahead nav.

The lookahead nav is rather primitive. If an upcoming turn is 90 degrees, it makes the 45 degree turn now.

I am using a global 'overallDirection' variable that should be updated or removed. We may only need this when we are following wall. Not sure...