Navigation no longer returns null.
Getting the overall direction in a different place so that if a bot is following a wall but then changes in the round it can get the overall direction.

The lookahead checks if the diagonal is navigable. If it is not it returns the original destiation.

I found another issue where a robot can go diagonal and then end up on a different obstacle which it cannot get off of. This may be solved by avoiding that obstacle or by resetting the distance when the bot starts to bug around the new obstacle.