Updated Bug class
Robots will randomly default left or right.

Added a hack where every fourth move a robot will see if it was here four moves ago.
This is to fix the problem where a robot goes in a loop and cannot get out.
The robot gets stuck because it can always move in the "check direction" but it is never
closer to the destination that when it started bugging. This happens because another unit
acts like a wall but then moves away.