Updated supply logic
First, there is a job for supplying the miners. If a certain number of miners have no supply, the job is needed. One drone takes the job.

Second, we will build a supply tower if we have more than 30 drone+miner units. To accomplish this I started to broadcast more unit counts.