Major overhaul to unit responsibilities
Orders have replaced Tactics. See issue #14.

Orders are only set by the HQ. They can have the value ON or OFF. They require no response that they are being done. This makes them different than jobs, which require acknowledgement from whoever picks up the job.

Orders are used to manage unit spawning, and they can also set a specific strategy, like "Pester with drones".

Unit spawning is also controlled by orders. The HQ will order to spawn a unit type. The building responsible for spawning will spawn and will be broadcasting the type.
The HQ can use the updated information about the unit count to decide to turn the order ON or OFF.

Previously, each building had its own logic about whether or not to produce more units. This is better for the HQ to do, because it is in a position to know more about the overall
strategy. Buildings are responsible for counting the units they produce, and if we have a lot of robots this can take 2 rounds to do.

The order that buildings are built is controlled with a queue. Building ordering is controlled with a queue. The HQ has an initial building list, but it can add more buildings later.
The beaver will build the building in the queue.