Added a build with delay method to BuildingQueue
which prevents a building from being added multiple times to the queue.
The scenario where I needed this is adding buildings dynamically. If we have a certain amount of bytecode or a certain number of units, we add a building to the queue.
The problem is that if this condition is true on round x, it is probably true on round x + 1. If we only want to add a single building to the queue, the delay works.
The delay is building type specific.

Another solution is to use the one time trigers, which I am using for the supply depots.

Spawner spawns before it broadcasts unit counts. I changed this because the broadcasting can take over 1800 bytecodes. When this happens the teamOre check can happen before the 2000 cutoff and the robot tries to spawn in the next round with insufficient ore.

This bot is destroying microLending on all 6 maps.