Orders have replaced jobs.
There are two different types of orders, priority orders and default orders.

Priority orders have a count of the max number of robots that should participate.
The management of orders is done through the message board.

The HQ must always set a default order per robot type. For priority orders, the HQ
only broadcasts an order if it is needed. There is no need to turn it ON/OFF.

Currently, only one priority order per unit is enabled. Having multiple priority orders
is possible, but would add some complexity when orders expire and robots are re-assigned.

In the bot now, the A team soldiers will pester the enemies miners for the first 500 turns.
Only 10 bots engage in this.