Soldiers and miners spread out.
Miners checks other miners it is close to. If therea are too many, it moves away from our HQ.
Default miner behavior is to more away from enemy HQ, so this results in spreading.

Soldiers check if they are standing next to too many soldiers. If they are, there get a direction
to move away.