Updated soldier defensive micro
I am hoping this will help with #19 because soldiers are no longer so tied to the miner radius. On voracious, the miner radius
was forcing the soldiers to move a certain distance from the hq, which put them in odd places. By having the soldiers move less and respond to enemies more,
I am hoping they will be better placed for early defense.

Towers are broadcasting "enemy spotted!". See issue #20. For now, this is separate from the distress signal logic.