More elaborate Drone micro
It all depends on how many enemies can shoot it, and how many of our teammates can shoot that set.

I fixed a mistake in the Helper.getDirection.. method. I was setting the value instead of using +=.

I added a BasicNavigation class to handle the tryMove method that I have been putting in so many classes.