Drone does not go where 2 miners/beavers can shoot
Safe bug was checking if the robot would be put into attack range of specific types of
units. This has the drone check if it is going to go to a map location that has more than one unit.
It will never outnumber itself, and it will only jump in if the unit is a beaver or drone.

This change increased the win percentage over justInTime from 60% to 72%.